third_party/blink/web_tests/platform/generic/external/wpt/encoding/legacy-mb-schinese/gbk/gbk-decoder.any.worker-expected.txt

This is a testharness.js-based test.
Found 82 tests; 1 PASS, 81 FAIL, 0 TIMEOUT, 0 NOTRUN.
FAIL gbk pointer: 6432 assert_equals: expected 8364 but got 59244
FAIL gbk pointer: 7533 assert_equals: expected 7743 but got 59335
PASS gbk pointer: 7536
FAIL gbk pointer: 7672 assert_equals: expected 12350 but got 59367
FAIL gbk pointer: 7673 assert_equals: expected 12272 but got 59368
FAIL gbk pointer: 7674 assert_equals: expected 12273 but got 59369
FAIL gbk pointer: 7675 assert_equals: expected 12274 but got 59370
FAIL gbk pointer: 7676 assert_equals: expected 12275 but got 59371
FAIL gbk pointer: 7677 assert_equals: expected 12276 but got 59372
FAIL gbk pointer: 7678 assert_equals: expected 12277 but got 59373
FAIL gbk pointer: 7679 assert_equals: expected 12278 but got 59374
FAIL gbk pointer: 7680 assert_equals: expected 12279 but got 59375
FAIL gbk pointer: 7681 assert_equals: expected 12280 but got 59376
FAIL gbk pointer: 7682 assert_equals: expected 12281 but got 59377
FAIL gbk pointer: 7683 assert_equals: expected 12282 but got 59378
FAIL gbk pointer: 7684 assert_equals: expected 12283 but got 59379
FAIL gbk pointer: 23766 assert_equals: expected 11905 but got 59413
FAIL gbk pointer: 23770 assert_equals: expected 11908 but got 59417
FAIL gbk pointer: 23771 assert_equals: expected 13427 but got 59418
FAIL gbk pointer: 23772 assert_equals: expected 13383 but got 59419
FAIL gbk pointer: 23773 assert_equals: expected 11912 but got 59420
FAIL gbk pointer: 23774 assert_equals: expected 11915 but got 59421
FAIL gbk pointer: 23776 assert_equals: expected 13726 but got 59423
FAIL gbk pointer: 23777 assert_equals: expected 13850 but got 59424
FAIL gbk pointer: 23778 assert_equals: expected 13838 but got 59425
FAIL gbk pointer: 23779 assert_equals: expected 11916 but got 59426
FAIL gbk pointer: 23780 assert_equals: expected 11927 but got 59427
FAIL gbk pointer: 23781 assert_equals: expected 14702 but got 59428
FAIL gbk pointer: 23782 assert_equals: expected 14616 but got 59429
FAIL gbk pointer: 23784 assert_equals: expected 14799 but got 59431
FAIL gbk pointer: 23785 assert_equals: expected 14815 but got 59432
FAIL gbk pointer: 23786 assert_equals: expected 14963 but got 59433
FAIL gbk pointer: 23787 assert_equals: expected 14800 but got 59434
FAIL gbk pointer: 23790 assert_equals: expected 15182 but got 59437
FAIL gbk pointer: 23791 assert_equals: expected 15470 but got 59438
FAIL gbk pointer: 23792 assert_equals: expected 15584 but got 59439
FAIL gbk pointer: 23793 assert_equals: expected 11943 but got 59440
FAIL gbk pointer: 23796 assert_equals: expected 11946 but got 59443
FAIL gbk pointer: 23797 assert_equals: expected 16470 but got 59444
FAIL gbk pointer: 23798 assert_equals: expected 16735 but got 59445
FAIL gbk pointer: 23799 assert_equals: expected 11950 but got 59446
FAIL gbk pointer: 23800 assert_equals: expected 17207 but got 59447
FAIL gbk pointer: 23801 assert_equals: expected 11955 but got 59448
FAIL gbk pointer: 23802 assert_equals: expected 11958 but got 59449
FAIL gbk pointer: 23803 assert_equals: expected 11959 but got 59450
FAIL gbk pointer: 23805 assert_equals: expected 17329 but got 59452
FAIL gbk pointer: 23806 assert_equals: expected 17324 but got 59453
FAIL gbk pointer: 23807 assert_equals: expected 11963 but got 59454
FAIL gbk pointer: 23808 assert_equals: expected 17373 but got 59455
FAIL gbk pointer: 23809 assert_equals: expected 17622 but got 59456
FAIL gbk pointer: 23810 assert_equals: expected 18017 but got 59457
FAIL gbk pointer: 23811 assert_equals: expected 17996 but got 59458
FAIL gbk pointer: 23813 assert_equals: expected 18211 but got 59460
FAIL gbk pointer: 23814 assert_equals: expected 18217 but got 59461
FAIL gbk pointer: 23815 assert_equals: expected 18300 but got 59462
FAIL gbk pointer: 23816 assert_equals: expected 18317 but got 59463
FAIL gbk pointer: 23817 assert_equals: expected 11978 but got 59464
FAIL gbk pointer: 23818 assert_equals: expected 18759 but got 59465
FAIL gbk pointer: 23819 assert_equals: expected 18810 but got 59466
FAIL gbk pointer: 23820 assert_equals: expected 18813 but got 59467
FAIL gbk pointer: 23821 assert_equals: expected 18818 but got 59468
FAIL gbk pointer: 23822 assert_equals: expected 18819 but got 59469
FAIL gbk pointer: 23823 assert_equals: expected 18821 but got 59470
FAIL gbk pointer: 23824 assert_equals: expected 18822 but got 59471
FAIL gbk pointer: 23825 assert_equals: expected 18847 but got 59472
FAIL gbk pointer: 23826 assert_equals: expected 18843 but got 59473
FAIL gbk pointer: 23827 assert_equals: expected 18871 but got 59474
FAIL gbk pointer: 23828 assert_equals: expected 18870 but got 59475
FAIL gbk pointer: 23831 assert_equals: expected 19619 but got 59478
FAIL gbk pointer: 23832 assert_equals: expected 19615 but got 59479
FAIL gbk pointer: 23833 assert_equals: expected 19616 but got 59480
FAIL gbk pointer: 23834 assert_equals: expected 19617 but got 59481
FAIL gbk pointer: 23835 assert_equals: expected 19575 but got 59482
FAIL gbk pointer: 23836 assert_equals: expected 19618 but got 59483
FAIL gbk pointer: 23837 assert_equals: expected 19731 but got 59484
FAIL gbk pointer: 23838 assert_equals: expected 19732 but got 59485
FAIL gbk pointer: 23839 assert_equals: expected 19733 but got 59486
FAIL gbk pointer: 23840 assert_equals: expected 19734 but got 59487
FAIL gbk pointer: 23841 assert_equals: expected 19735 but got 59488
FAIL gbk pointer: 23842 assert_equals: expected 19736 but got 59489
FAIL gbk pointer: 23843 assert_equals: expected 19737 but got 59490
FAIL gbk pointer: 23844 assert_equals: expected 19886 but got 59491
Harness: the test ran to completion.
